Default bubbled message to the bubbleBuilder (#553)
* Update message.dart

Changes to the bubbleBuilder.  It now creates the default message, and sends it to the bubbleBuilder(). This way people can just spit back message as they see fit or wrap the message with their own widgets.

* Update chat.dart

Changed the definition for the bubbleBuilder.  It now has a defaultBubbleMessage which is the original message.